Arma2 OA Auto Steam EXE Changer Changed beta installer code slightly
Added feature so that when you have "Using Steam" and "Update steam exe" enabled it checks the main exe version with the beta exe just before launching game and backups / updates the EXE if needed. This allows people to perform manual beta installations and for those who enable this option but havent run a beta install.